Fix potential problem in 'file copy' and 'file delete'
The full syntax of these commands is like this
   file delete ?-force? ?--? pathname ?pathname ... ?
These commands accept flags (currently one -force). But that means if the file
you want to delete/copy starts with a '-', it will be seen as a (probably
unknown) flag. Fixed by adding the '--' marker (after this marker all arguments
are seen as files, even if they start with a '-').

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12031 7512d703-071c-0410-b5c2-d2bdd895b7d1