Lots of small layout changes and expr-tweaks, there shouldn't be any functional changes
- Don't put a space right after '{' or before '}'
- Always put {..} around the expression in an 'expr' and 'if' command.
  So replace stuff like:
     expr [get_value] * 2   -->    expr {[get_value] * 2}
     if [some_condition]    -->    if {[some_condition]}
  This is more ugly, but it has a real benefit. I'll try to explain below:

These two commands do NOT always do the same:
   expr  $a * 3
   expr {$a * 3}
For example:
   set x 1
   set a {$x + 2}
   expr  $a * 3   --> result = 7    (not 9 as you may have expected)
   expr {$a * 3}  --> ERROR: can't use non-numeric string as operand of "*"

The first example does two rounds of substitution (once while calculating the
arguments that must be passed to the expr command, and once while actually
evaluating the expression). So the expr command sees as input '$x + 2 * 3',
while evaluating it substitutes '$x' for the numeric value '1', and '1 + 2 * 3'
finally results in '7'.

In the second example, because of the {..}, there is no first round of
substitution, so the expr command sees '$a * 3'. It will try to substitute
'$a' for a numeric value, but that conversion fails.

So an 'expr' command with or without braces can theoretically behave different,
though in practice these situations rarely occur (in the example above you most
likely want to write 'set a [expr {$x + 2}]', that will also result in the (for
me) more expected value of '9'). Still the Tcl interpreter has to account for
the possibility. This can make a _big_ difference in the evaluation speed of
expressions with or without braces around them:

   time {expr $x + 1}   1000000 --> 1.566082 microseconds per iteration
   time {expr {$x + 1}} 1000000 --> 0.172188 microseconds per iteration

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12042 7512d703-071c-0410-b5c2-d2bdd895b7d1