Ran a static analyzer on all Tcl scripts and fixed warnings
http://catless.ncl.ac.uk/Programs/Frink/

There were only two types of warnings left:
 - switch statement without default case
 - switch statement without '--'
(There are also false-positive warnings about vars being used before set.)

For example:
    switch $a { .. }  -->  switch -- $a { .. }
The '--' prevents the value '$a' from being interpreted as a flag. Even if you
know for sure the value cannot start with a '-' character (e.g. because it is a
byte you 'peek'ed from memory, so a numeric value between 0-255), it is still
useful to add '--', because with '--' the switch runs faster.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12043 7512d703-071c-0410-b5c2-d2bdd895b7d1