Show different output for the two key joystick descriptions.
This required changing all pluggables... Also cleaned up some of the
names. There is still a lot to do, because we should use the
descriptions only as a display value for humans and not as a manual that
explains what they do, just like the connectors.
Also, the description of the cassette player is also used as the
description of its audio device part, and that is used as a way to
identify this audio device in wxCatapult... I didn't want to change that
now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12045 7512d703-071c-0410-b5c2-d2bdd895b7d1