Only show Joy Tap and Ninja Tap ports when available.
Before this patch, the Joy Tap and Ninja Tap ports were always shown,
even when these devices weren't plugged in the MSX. This is confusing
and even annoying, because it clutters menus with connectors.

When unplugging a tap device, all pluggables in them are unplugged,
otherwise it would not be possible anymore to use those pluggables
elsewhere (you couldn't even unplug them). The only disadvantage about
this is that if you made a mistake and put the tap device in the wrong
port, and already plugged it full of pluggables, you will have to do
that again after plugging it in the right port. But I think that's well
worth it :)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12047 7512d703-071c-0410-b5c2-d2bdd895b7d1