YM2413Burczynski refactoring: removed 'kcodeScaled' member variable
This member variable is easily calculated from other members. It was also used
to avoid some recalculations when the new 'kcodeScaled' value is the same as
the old value. Though these calculations are relatively cheap, and this value
changes relatively infrequent. So I don't think it's worth it. 'kcodeScaled'
also doesn't need to be serialized anymore.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12057 7512d703-071c-0410-b5c2-d2bdd895b7d1