Implemented SETTLE (or DUMP) phase for YM2413Okazaki
I have no idea if this implementation is correct. It's based on the extremely
limited information in the ym2413 application manual and some comments from
the Burczynski core.

As a test, in MoonBlaster I made a custom instrument with a very slow decay
rate and played the same note successively. I *believe* the sound generation
after this patch is closer to the real hardware than without this patch (more
easy to distinguish the successive notes). But this needs more testing by
someone who's better than me at this stuff.

... also compared it now with the output of a real YM2413 (thanks Meits). After
some small tweaks, the output of a real YM2413 and openMSX emulation look very
similar.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12071 7512d703-071c-0410-b5c2-d2bdd895b7d1