Optimized PixelOperations for Dingoo
The pixel format on Dingoo is fixed. We can gain some speed when the pixel
format related constants are also compile-time constant. The speed gain will be
much more noticeable when in some future patch horizontal stretch is also
implemented for the SDL renderer.

This refactoring also allowed to make a small optimization for 32bpp pixel
formats: for those the PixelOperations class doesn't need to store the
'blendMask' member variable. (For dingoo, the PixelOperations class doesn't
have any state anymore). So constructing and copying this class is now a tiny
bit faster.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12117 7512d703-071c-0410-b5c2-d2bdd895b7d1