Refactor output side of the SDL scalers
Before this patch, SDL scalers would output directly to a OutputSurface. Now
they output to an abstract ScalerOutput class. This allows to (in follow up
patches) still process this output. E.g. super-impose output with a laserdisc
frame, or apply horizontal stretch on each line. The current patch only
refactors the scalers. It doesn't yet add any post-processing effects.

Before this patch a scaler would request a Pixel buffer via getLinePtrDirect(),
then it simply filled that buffer with scaled pixels. Now it still has to
request a buffer (but via acquireLine()) but then it also has to signal when
it's done using that buffer via a call to releaseLine(). This allows for
OutputSurface subclasses to post-process each line as soon as it's scaled. So
this requires (potentially) less buffer space and it also can work on data that
is still warm in the CPU cache.

This patch is rather big in terms of changed number of source code lines,
though most changes are relatively simple. Here's a summary of the changes:
 - Changed 'dst' parameter in a lot of the scaler methods from type
   'OutputSurface' to 'ScalerOutput<Pixel>'. (Not yet in the scaleImage()
   method, this will be done in the next patch, when we implement super-impose
   as a post-processing step).
 - In the implementation of all scaler methods, replaced calls to
   getLinePtrDirect() with a pair of calls to acquireLine() and releaseLine().
 - Replaced all uses of MemoryOps::MemSet from the scaler implementations with
   calls to fillLine(). This is an optimization, future ScalerOutput subclasses
   may perform post-processing more efficient when they know the whole line has
   a constant color. A call to fillLine() replaces the pair of acquireLine(),
   releaseLine() calls.
 - Before this patch some scaler implementations called getWidth() or
   getHeight() more than once. Those methods are now virtual, so it may be
   better to only call them once and cache the result in a local variable
 - Added DirectScalerOutput class. For now this is the only concrete
   implementation of the ScalerOutput interface. In acquireLine() this class
   returns a direct buffer from the underlying OutputSurface. In releaseLine()
   this class does nothing. So the effect is exactly the same as before this
   patch.
 - Removed all calls to OutputSurface::lock() from all scaler implementations,
   see below.

In the Scaler1 class we had some GP2X specific code that used the hardware
blitter for line-copies (scale a line 1-on-1) and line-fills (border line with
a single color). For now that code is disabled because it didn't fit well in
the new structure. Maybe in the future we can re-enable it, though the GP2X
port isn't actively maintained anymore (performance wasn't good enough).
Because of this change scaler implementation no longer have to (un)lock the
OutputSurface.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12123 7512d703-071c-0410-b5c2-d2bdd895b7d1