Refactor calculation of slot_on_flag
Instead of keeping only one boolean status and keeping it in sync with both
the FM and DRUM keyon register bits, we now keep two bits that are each in sync
with respectively the FM and DRUM register bits. The 'logical' slot_on status
is the OR of these two bits (so we can simply pack the two bits in an int and
compare that int with zero).

This is also how the Burczynski core and the YMF262 core implement this. I also
verified on a real YM2413 that the logical slot_on state is the OR of the FM
keyon bit and the DRUM keyon bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12140 7512d703-071c-0410-b5c2-d2bdd895b7d1