Fixed File::getOriginalName() for .gz files
Some .gz files don't store the original filename (= the filename before the
file was compressed). For example our .omr files are immediately generated in
compressed format, so these don't have an original filename.

For such files, the method File::getOriginalName() used to return an empty
string. Now we return the current filename instead.

This fixes the loading of .omr files on the command line (this patch is not yet
committed, but Manuel is working on it). Filename extensions are recognized
based on the original filename. But of course this goes wrong when it has to
work on an empty string.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12142 7512d703-071c-0410-b5c2-d2bdd895b7d1