Removed too strict assert
Maybe in the past the condition in the assert was required, but I don't see
a reason for it (anymore).

This assert would trigger when you tried to load a replay when reverse was not
(yet) enabled. This happened in Manuels work-in-progress "load replay from the
commandline'-patch. But it could also be triggered from the openmsx console.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12143 7512d703-071c-0410-b5c2-d2bdd895b7d1