Fixed execution of VDP command closely followed by VDP screen mode switch
There was a bug in the VDPCmdEngine subsystem sync()-ing. Because of this the
last part of a command could get executed in the new screen mode, even if the
command would already be finished by the time the new screen mode is selected.

This bug was visible in e.g. 'Psycho World'. The 3x3 power-up matrix in the
status region at the bottom should initially contain 7 empty squares. But
because of this bug, it already contained the 9 power-up icons. The VDP
commands that would erase those squares were not properly executed.

Initially we thought 'Psycho World' actually tried to execute VDP commands
while switching to character mode. But (trying to) fix the behavior of VDP
commands in char mode didn't solve the problem. Now we know the bug was
something else, but the emulated behavior for VDP commands in character mode is
still not correct. TODO:
  Initial tests indicate that the command continues as usual in character mode
  except that writes to VRAM have no effect. But this also needs more tests
  on a real machine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12173 7512d703-071c-0410-b5c2-d2bdd895b7d1