Added better emulation of the Y8950 BUF_RDY status bit
Before this patch, the BUF_RDY status bit was always read as '1' (at least when
it was not masked). Most of the time this is good enough. But in programs like
'NOP real motion', this bit was used to control the rate at which the CPU
should feed new data to the ADPCM unit (most other programs play adpcm samples
via the sample-ram). But when this bit was always '1', the data was fed too
fast and (because of buffer overflows??) the program crashed.

Fixing this required better emulation of the BUF_RDY bit. In my first attempt I
missed some corner cases and that caused new bugs (e.g. hang when 'NOP unknown
reality' was checking the amount of available sample ram or uploading new
samples). I checked a number of programs and they all (still/again) seem to
work fine. I just wanted to say that it's not impossible that this patch causes
regressions.

The sound quality in 'NOP real motion' is still very bad (while on a real
machine it's quite good). But that's something for a later patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12179 7512d703-071c-0410-b5c2-d2bdd895b7d1