Improve handing of re(g)zipped files (e.g. OMR).
First look at the extension of the file as given on the command line,
and if that doesn't work (e.g. .gz or .zip file), look at the original
file name.
Before this patch, re(g)zipped OMR files wouldn't be recognized, because
only the original file name (inside the gzipped file) would be used,
which is not OMR at all, usually.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12195 7512d703-071c-0410-b5c2-d2bdd895b7d1