Increase the resolution of the pre-calculated impulse-response in BlipBuffer
In this BASIC test:
  SOUND 0,0
  SOUND 1,0
  SOUND 8,15
You could here same noise when using the blip-resampler (not with the hq
resampler). After this patch the noise with blip is a _lot_ less. (AFAIK it's
not possible to completely remove it with the current design of BlipBuffer).

The original value for BLIP_PHASE_BITS was 6, this patch changes it to 10. Our
tests showed that for values 9 and above the sound quality doesn't improve
anymore. But just to be on the save side we took one more (it might be the case
that for other tests there still is a minor difference between 9 and 10).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12199 7512d703-071c-0410-b5c2-d2bdd895b7d1