Make our ARM inline asm routines compile in Thumb2 mode
Modern ARM distributions (e.g. Ubuntu 11.04) by default generate code for
ARM-Thumb2 mode. But before this patch, our code contained some ARM inline asm
routines that are not valid Thumb2 instructions. Now the code should compile
fine in both ARM- and Thumb2-mode (but not Thumb1).

FYI:
 ARM-mode:    All instructions are 32-bit.
 Thumb1-mode: All instructions are 16-bit, (of course) these can only
              represent a subset of the full (32-bit) ARM instructions.
 Thumb2-mode: Mix of 16-bit and 32-bit instructions. Many (but not all)
              ARM instructions can directly be represented as a single Thumb2
              instruction. The most common instructions only take 16-bit, the
              others are 32-bit (of course with some limitations).

We could have 'fixed' the compilation errors by simply forcing ARM-mode (gcc
compiler switch '-marm'). But that would completely eliminate the code-size
advantage of Thumb2 mode (and because of better cache locality this often also
implies a speed-advantage). (There are also a few ARM cores that only support
Thumb2-mode (no ARM-mode), OTOH those cores are often also not suitable to be
used as a general purpose application CPU.)

Because most ARM instructions have a direct mapping to a single Thumb2
instruction, it wasn't that hard to rewrite our code so that it compiles in
both ARM and Thumb2 mode. Making it also compile in Thumb1 mode is much harder
and it would also degrade performance when that same code is compiled to ARM or
Thumb2 code.

These are the most important changes I had to make:

- In ARM-mode there's an instruction:
     ORR Rd,Rn,Rm,LSL Rs        // equivalent to:  Rd = Rn | (Rm << Rs)
  In thumb2-mode these 'inline shifts' are only possible with an immediate
  shift count (not a variable count from another register as in this case). As
  a solution I've split this instruction in two. For ARM-mode this introduces a
  small inefficiency, but that's a small price compared to having to maintain
  two separate versions.

- In ARM-mode each instruction can be conditionally executed. In Thumb2 there's
  a 'IT..' instruction that conditionally executes the next (1-4) instructions.
  Though luckily the same asm syntax can be used for both modes (when you use
  the 'IT..' instruction in your asm code, it either expands to a real thumb2
  instruction, or it only influences the condition bits in the next
  instructions).

- In some cases gcc complains about the use of register 'r7' (thumb2 mode) or
  register 'r11' (arm mode) (not clear to me when/why exactly, though it's
  related to the use of this same register as a frame-pointer-register). As a
  workaround I've rewritten the code to avoid using either of these two
  registers.

- Some tweaks to allow to use short (16-bit) instead of long (32-bit)
  instructions. For example, in ARM and thumb2 mode, for most instructions,
  there's a variant that does and one that does not update the status flags.
  Short thumb2 instructions often only have the variant that does update the
  flags. So when the value of the status flags doesn't matter it's sometimes
  better to choose the variant that does update the flags.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12202 7512d703-071c-0410-b5c2-d2bdd895b7d1