Fixed drum-bug regression in YM2413Okazaki
The bug could be triggered like this:
 Play the attached .MBM file in this bug report:
     http://sourceforge.net/tracker/?func=detail&atid=421861&aid=3380296&group_id=38274
 If you look at a deeper level, playing this song (only) writes, in an
 alternating way, the values 0x01 and 0x21 to YM2413 register 0x0E.

The regression was introduced in revision 12140. This was a refactoring to
bring the Okazaki and Burczynski code closer together. It was also based on
measurements on a real YM2413. Though there was still one part in the Okazaki
code that was left unchanged (not changed into the Burczynski behavior). That
code was marked with a "TODO compare this difference between Okazaki and
Burczynski with real YM2413". The bug was caused by the combination of the
partly changed/partly unchanged code.

This patch also changes that previously unchanged part into the Burczynski
behavior. That behavior indeed seems more likely to be the correct behavior (to
me this *seems* easier to implement in HW). And this bug confirms it. As a
bonus this fix also simplifies the (software) implementation of Okazaki a bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12211 7512d703-071c-0410-b5c2-d2bdd895b7d1