Fixed potential out-of-memory problem when using step_back
To reproduce:
- Start openmsx using a 'big' machine. E.g. a turbor with moonsound and v9990
  extensions. (A turbor in R800 mode makes the bug _much_ easier to trigger,
  see below).
- Execute the 'step_back' a lot of times (sometimes it happens after a few
  times, other times I had to step_back dozens of times).
  Or alternatively, execute this command:
     for {set i 0} {$i < 100} {incr i} {reverse goto 1}
  (this has no practical use, but it easily triggers the problem).
After this you will see a _huge_ increase in memory usage. Sometimes even so
big that the whole OS becomes unresponsive (for some time). Though if the OS
recovers, openmsx keeps on working normally and if you run a tool like
valgrind, it won't find any memory leaks (openmsx at one point allocates a huge
amount of memory but later properly frees it all).

The problem was the following: the implementation of the 'step_back' script
calls 'reverse goto' multiple times, each time going back in time one more cpu
cycle. Each 'step' creates a new MSXMotherBoard object, but the old objects only
get deleted once the whole Tcl scripts is finished.

The reason for this delayed deletion is explained in (the ChangeLog of)
revision 7897. Very briefly it's because we can't delete a MSXMotherBoard
immediately when we switch from one board to another because the command that
initiated the switch is often running on behalf of the old board (so there are
still call frames active that may refer to data of the old board).

In this specific case (reverse goto) it was possible to bypass the delayed
deletion mechanism by carefully ordering the operations so that none of the
still active old MSXMotherBoard code still accesses any (already deleted)
member variables.

So this patch fixed the huge memory consumption problem. But it does not yet
explain why the problem was triggered in the first place. I would have expected
that stepping back a single instruction shouldn't take more than a couple of
steps:
 - On Z80 the step_back script can take steps of 5 cycles (the shortest
   instruction takes 5 cycles). The longest instruction ('EX (SP),IX' I guess?)
   takes 25 cycles. So 5 steps should always be enough. Fine.
 - On R800 the shortest instruction takes only 1 cycle, so we can also only
   takes steps back of 1 cycle. Though the variation in the length of the
   instructions is much bigger. Most instructions finish in less than 10 cycles
   even taking into account the penalties for page-breaks or access to
   non-internal RAM. Occasionally there may be a 'MULUW' instruction that takes
   36 cycles (without penalties). But there's also I/O to the VDP which will
   stall the CPU up to 62 cycles. And then there's the refresh stuff every 210
   cycles which will stall the cpu for 25 cycles. So all combined the longest
   instruction can take up to 87(!) cycles. And this is a lot to step back
   cycle per cycle.
So while the memory problem is fixed, using step_back in R800 mode may still be
too slow occasionally. Though fixing that is something for a later patch
(probably requires a different step_back algorithm).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12219 7512d703-071c-0410-b5c2-d2bdd895b7d1