Fixed crash when clicking in reversebar
This regression was recently introduced in revision 12219. The scenario that
caused the crash went like this:
 - User clicks on the reversebar
 - A mouse-button-down event gets distributed to all the currently subscribed
   listeners (a copy of this list of listeners is made)
 - One of the listeners for this event is the OSDGUI layer. Via some Tcl
   scripts, a click in the reversebar will replace the current MSXMotherBoard
   with a new one at a different point in the timeline.
 - Another listener for the mouse event is the per MSXMotherBoard EventDelay
   class, but that one has just been deleted!

To fix this problem, right before actually delivering an event to a listener, I
now check if this listener is still actually subscribed. Note that it's (still)
required to make a copy of the list of listeners because iterating over a
changing STL collection is also undefined behavior (the collection changes when
a listener (un)subscribes).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12222 7512d703-071c-0410-b5c2-d2bdd895b7d1