Speedup step_back [1/3]: speedup 'reverse goto' for small forward steps
This is a preparation to speedup step_back (at least speedup the worst case
execution time). See commit 12219 for a description of why/when step_back
is still very slow in some cases (R800 mode).

Before this patch 'reverse goto' always took the oldest snapshot from history
that was still before the target time and then ran this snapshot (in
fast-forward mode) till the requested target time was reached.

However it may be the case that the current MSXMotherBoard is actually closer
to the target time than any of the available snapshots. In that case it's
faster to simply fast-forward from the current state instead of first replacing
it with a snapshot.

(In a follow-up patch I'll rewrite the step_back script to take advantage of
this optimization).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12223 7512d703-071c-0410-b5c2-d2bdd895b7d1