Speedup step_back [2/3]: added -novideo flag to 'reverse goto'
Internally 'reverse goto' first goes to the point in time that is 2 video
frames before the requested time. And only then it goes to the actual target
time. This makes sure the video output is properly rendered when the final
target time is reached.

However this behaviour also defeats the small-forward-step optimization that
was implemented in the previous patch (at least when the step is smaller than
the duration of 2 frames). This patch extends 'reverse goto' to (optionally)
accept a '-novideo' flag. When this flag is passed, 'reverse goto' will go
immediately to the target time (though of course the correct video output is no
longer guaranteed directly after the goto).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12224 7512d703-071c-0410-b5c2-d2bdd895b7d1