Fixed hang on boot with Panasonic_FS-CA1 extension
This was a regression introduced in revision 12179. To reproduce, simply
(try to) boot to BASIC with the Panasonic_FS-CA1 extension inserted.

The problem was the following:
 - The FS-CA1 boot ROM at one point writes the value 0x60 to Y8950 register R#7
   (the command to start transfer from CPU to sample-RAM, in other words upload
   ADPCM sample data).
 - Then it wait in a loop till the BUF_RDY status bit (bit 3) is set. Though
   before this patch, this never happened. So this results in an endless loop
   (=hang).

Uploading data to the sample-RAM seems like very basic functionality. The
reason why this bug was not detected earlier is the following:
 - Before revision 12179 we cheated and always had bit BUF_RDY set to '1'.
 - Even after rev 12179 much software kept working because after the first data
   byte is written to the sample-RAM (via register R#15), we _did_ set the
   BUF_RDY bit. And if you look in the Y8950 datasheet, the recommended
   algorithm to upload data does indeed first write one data byte before
   checking the BUF_RDY bit.

The hardest part of fixing this bug was figuring out what went wrong. Once that
was known, the fix was easy: set the BUF_RDY bit as soon as the 'start
transfer' command is given.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12226 7512d703-071c-0410-b5c2-d2bdd895b7d1