Refactored Resample -> ResampledSoundDevice
Refactored the class 'Resample' into the class 'ResampledSoundDevice'. This
removes quite a bit of code duplication between the different SoundDevices that
resample their output stream (almost all SoundDevices). This patch should not
cause any functional changes. As a bonus it also allowed to reduce the #include
dependencies of most SoundDevices.

This patch on its own is a nice cleanup. Though the main motivation for it is
as preparation for a later patch that will change how resampling is done
(synced on EmuTime instead of trying to estimate the actual emulation speed and
trying to compensate for differences). This patch (_very_ roughly) separates
the per-SoundDevice changes from the per-Resampler changes. Also this part is
low risk, while the next part will make fundamental changes. Splitting it
allows for better testing or at least better bug localization.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12229 7512d703-071c-0410-b5c2-d2bdd895b7d1