Fixed initialization bug in YM2151
To reproduce start openMSX with:
  openmsx -machine Yamaha_CX5M YRM104.rom musicmacro2.cas
then in BASIC execute
  CLOAD
  RUN
and it will crash a few seconds after finishing loading from tape.

The crash was caused by accessing an array with a negative index. Further
investigation showed that 'normally' this index should never be negative, (the
member variable YM2151Operator::kc_i should always be >= 768) except right
after a reset when registers 0x28-0x2F have not been written yet. Actually the
implementation of the reset() method does internally write to those registers
(so the variable 'kc_i' should have been given a proper value) except that the
implementation of writeReg() skips some stuff when the newly written value is
the same as the old one (this is a perfectly fine optimization, except for
initializing the values for the very first time).

I've also compared our YM2151 code with the code from MAME, and there in the
ym2151_reset_chip() function, the variable 'kc_i' is explicitly initialized.
For some reason that line is missing in the corresponding openMSX
YM2151::reset() method. I've checked SVN history and it's always been like this
in openMSX. In MAME it's like that since at least 2007/12/17 (MAME SVN history
doesn't go back further). So maybe this is bug that has been fixed in MAME
since the time we ported the code or maybe we made a mistake in the initial
port. TODO It may be worth comparing both code bases for other bug fixes and/or
mistakes.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12233 7512d703-071c-0410-b5c2-d2bdd895b7d1