Sound was muted when taking small forward 'reverse goto' steps
Could easily be triggered by pressing 'PgDown' or by clicking in the reverse
bar close to the current position.

This bug was introduced in SVN revision 12223. As an optimization this revision
sometimes reuses the current board instead of always creating a new one. Though
in case we reuse the old board it's not OK to leave the old board muted. See
comments in code for more details.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12236 7512d703-071c-0410-b5c2-d2bdd895b7d1