Refactor sound device resampling, part 1
Before this patch openMSX had very bad sound quality for samples that are
played in a Z80 controlled loop (meaning the Z80 is responsible for OUT-ing
each individual sample at the correct moment in time (so not e.g. Y8950 ADPCM
samples that, once started, play automatically without help from the Z80). The
reason is the following:

It's easier explain with an example. Suppose the openMSX host sample rate is
20kHz and the Z80 is playing a sample of 15kHz on a sound device with a native
sample rate of 30kHz. This means the host expects to see a sample every 50us
(at times 0, 50, 150, 200, ...) while the software running on the Z80 should
produce a new sample approx every 66us (0, 66, 133, 200, 266, ...) and the
sound device will produce a sample every 33us (0, 33, 67, 100, 133, ...). So
what should happen is that the sound device each time produces two successive
samples with the same value. Those emu samples are then resampled to the
correct host sample rate (band-limited resampling, so without introducing
aliasing effects).

Each time the Z80 does an OUT to the sound device (e.g. to rapidly change the
volume of some output channel), the openMSX sound system will 'sync' the host
and the emulated audio stream. In the (old) syncing mechanism there was a
weakness. This is what happened:

 A1) Z80 writes a new value at time 66us, the audio stream is synced (right
     before doing the actual OUT)
 A2) We figure out 1 host sample has passed (at time 50us)
 A3) To resample N emu-samples to M host samples, we need N = 3/2 * M (for
     the numbers in this example). So in this case N = 1.5, rounding up gives
     N=2. So the sound device is asked to to produce (emulate) 2 samples.
 A4) These two samples are resampled to 1 host sample (the resampler is smart
     enough to know to 2nd sample is actually only half a sample)
 B1) At time 133us, Z80 writes a new value
 B2) Again 1 host sample has passed (at time 100us)
 B3) In the previous step we actually only needed 1.5 emu samples, but we
     did produce 2, so this time it's enough to only produce 1 new emu sample
 B4) The half sample from the previous sync is combined with the 1 new sample
     and resampled to 1 host sample
 C1) At time 200us Z80 writes a new value
 C2) Now 2 host samples have passed (at time 150us and 200us)
 C3) To produce two host samples, we need 3 emu samples
 C4) These 3 emu samples are resampled to 2 host samples

Now let's look at what happened from the emulated sound device point of view:
The device was asked to produce 2 samples (then change the volume), then
produce only 1 sample (change volume) and then 3 samples. So graphically:
     -----    --------
    |     |  |        |
   -       --          -
While ideally, like we said earlier, the result should be:
     -----       -----
    |     |     |     |
   -       -----       -
This difference is clearly audible, it results in bad audio quality.

The underlying problem of the above 'sync' implementation is that it first
rounds the current time to an integer number of host samples and only then
calculates how many emu samples are needed to produce that many host samples.
Instead we should look at the previous and the current sync time and based on
this calculate BOTH how many emu AND how many host samples have passed. If you
run the above example through this new sync algorithm you'll see that it
produces the correct waveform.

There is one more complication. The Z80 writes to an OUT port at a specific
moment in emutime (thus also audio stream syncs happen at a specific emutime).
So we can easily figure out how much emutime has passed between 2 syncs. And
this allows to calculate how many emu samples have passed since the last sync.
But to calculate how many host samples should be produced, we need to know how
much realtime has passed.

The easiest solution is to assume realtime and emutime pass at the same rate.
And that is what this patch implements. A follow-up patch will take the 'speed'
setting into account, e.g. with speed=200 emutime will pass twice as fast as
realtime. Another complication is the 'throttle' setting, when throttle=off
there's no fixed ratio between emu and realtime.

Before this patch we had a mechanism that tried to dynamically measure the
ratio between emu and realtime. If we noticed the sound buffer is getting empty
(near buffer under-run) we start producing emu samples at a slightly higher
rate, if the buffer is too full (buffer over-run) the rate is decreased. It's
not clear whether this mechanism was useful or not. This mechanism was only
implemented in the SDL sound driver, the DirectX and the libao drivers don't
have it and they also work well. Other (msx) emulators I've looked at don't
have such a mechanism and they also work well. So probably it's not worth the
complexity. And the complexity would have increased when trying to adapt this
mechanism to the new audio-stream-sync stuff.

You may notice the removal of this mechanism when the host is not fast enough
(e.g. very high speed setting (or throttle=off) or a slow host machine). This
will result in buffer under-runs. It is still possible to compensate for buffer
overruns, see one of the follow-up patches.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12241 7512d703-071c-0410-b5c2-d2bdd895b7d1