Fix buffer overrun problem in SDLSoundDriver
In one of the previous patches, the mechanism in the SDLSoundDriver that tried
to keep the sound buffer not too empty and not too full was removed (see that
commit for details). Though now a buffer overrun does occasionally happen. But
there is a simple solution: if there's no space available in the sound buffer
to put new samples, then wait (=sleep) till the space becomes available. This
is also what the libao sound driver does (internally in the lib).

Though this approach does have a problem:
 - throttle=off has no effect anymore, because now the sounddriver will
   effectively throttle emulation to realtime speed
 - The time spend sleeping in the SDLSoundDriver is see by the RealTime class
   as 'active' time, while it is actually 'idle' time. This sometimes causes
   the RealTime class to (advise the video subsystem to) skip frames to keep
   up with realtime.
Both these problems will be fixed in the next patch. The 'libao' sounddriver
has the same problems (though not easily fixable).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12244 7512d703-071c-0410-b5c2-d2bdd895b7d1