Fixed bug in serialization framework: serialization of 'char' datatype
The types 'signed char' and 'unsigned char' were already serialized
numerically, but the type 'char' was serialized as a single character. At the
very least this can be confusing: the character '0' does not represent the
numerical value 0, but 48 (ASCII value for '0'). And for some values it may
even be wrong: I've verified that a 0-char doesn't end up in the output XML
(though on reading you again get the value 0). I'm not sure that whitespace
characters survive a save/load cycle. There may also be issues with chars > 128
interfering with utf-8 encoding.

The bug was caused by a combination of very subtle c++ (mis)features:
 - The types 'char', 'signed char' and 'unsigned char' are 3 distinct types in
   c/c++. For other types there are only two variants e.g. 'int' and 'signed
   int' is the same type. IOW it's possible to overload a function on 'char',
   'signed char' and 'unsigned char'. But overloading on 'int' and 'signed int'
   is not possible.
 - In the class XMLOutputArchive there's a method save() that has several
   overloads. Two of these overloads are
     void save(unsigned char);
     void save(signed   char);
   Normally if you call this method with a 'char' type, overload-resolution
   will call either the first or the second variant (which depends on the
   compiler/platform, the signedness of the 'char' type is implementation
   defined). However there's also a templatized overload:
     template<typename T> void save(const T& t);
   Overload resolution now figures out that when T=char, this 3rd variant is an
   even better match than the previous two.
 - The first two variants were explicitly added to avoid saving (un)signed char
   as a single character, but the 'char' type slipped through the net.
(XMLInutArchive::load() has the same problem)

A simple fix is to also overload XMLOutputArchive::save() on char. This works
(and does patch adds exactly this overload), but it also breaks backwards
compatibility.

Luckily only the YMF278Slot class was affected by this bug. To keep backwards
compatibility, I've added a new serializeChar() method in the serialization
framework. This method behaves as serialize(char) before this bug fix. The
YMF278Slot class now uses this to load older versions. Saving is done in the
new way (save char numerically).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12250 7512d703-071c-0410-b5c2-d2bdd895b7d1