Made de-serialization of bool types more robust
Currently bool types are serialized as the strings "true" or "false" (at least
to XML, in memory they only take a single byte). Before this patch de-serialization
of a bool type also only recognized these two strings. This patch makes it also
recognize the strings "0" and "1".

The reason for this change is the following: suppose some class has a member
variable of type int/char and at a later point this type is changed to bool
(e.g. because the code was ported from C which doesn't have a bool type and the
variable can only logically assume the values 0 and 1). Before this patch it
was required to update the serialize() method of this class to keep it
backwards compatible with old savesates. With this patch the old savestate can
still be read without extra changes. Note that changing from bool back to
int/char is still not backwards compatible.

I added this patch because in the YMF278 class there are a few variables that
could be changed from char->bool (I didn't commit this change yet). At first I
didn't realize that changing this would break bw-compat. I might be possible
that in the past/future I also forgot about this.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12251 7512d703-071c-0410-b5c2-d2bdd895b7d1